# executor: support spill intermediate data for unparalleled hash agg