<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #xxx <!-- REMOVE this line if no issue to close -->

Problem Summary: 

### What is changed and how it works?

Proposal: [Design](https://github.com/pingcap/tidb/pull/25792) <!-- REMOVE this line if not applicable -->

What's Changed:
Based PR https://github.com/pingcap/tidb/pull/25820, introduce soft limit.
Support spilling intermediate date for unparalleled hashAgg.

How it Works:
1. When the memory usage is higher than the quota, switch the HashAgg executor to spill-mode.
2. When HashAgg is in spill-mode, keep the tuple in the hashMap no longer growing.
 a. If the key exists in the Map, aggreagte the result.
 b. If the key doesn't exist in the Map, spill the data to disk.
3. After all data have been processed,  output the aggreagte result in the Map, clear the Map. Then read the spilling data from disk, repeat the Step1-Step3 until all data have been aggregated.  

### Related changes

- PR to update `pingcap/docs`/`pingcap/docs-cn`:

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Integration test
- Manual test (add detailed scripts or steps below)
Set AggregateConcurrency to 1 and run all correctness test in tidb repo.

Side effects

- Performance regression
    - Consumes more CPU
    - Consumes more MEM

### Release note <!-- bugfixes or new feature need a release note -->

- Support spill intermediate data for unparalleled hash agg<!-- Please write a release note here to describe the change you made when it is released to the users of TiDB. If your PR doesn't involve any change to TiDB(like test enhancements, RFC proposals...), you can write `No release note`. -->