@wshwsh12 2021-06-28
### Simple Benchmark
cpu: AMD Ryzen 7 3700X 8-Core Processor   
```
Old Implement:
BenchmarkHashAggRows
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:1,_sorted:true)
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:1,_sorted:true)-16         	      79	  23354686 ns/op
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:1,_sorted:true)
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:1,_sorted:true)-16        	       7	 150919829 ns/op
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:1,_sorted:true)
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:1,_sorted:true)-16       	       1	1315785257 ns/op

New Implement:
BenchmarkHashAggRows
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:1,_sorted:true)
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:1,_sorted:true)-16         	      52	  20963917 ns/op
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:1,_sorted:true)
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:1,_sorted:true)-16        	       7	 145583208 ns/op
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:1,_sorted:true)
BenchmarkHashAggRows/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:1,_sorted:true)-16       	       1	1327683063 ns/op
```
### Memory usage test
Workload: tpch-sf=3g  1tidb 1tikv 1pd 
Config: oom-action = "log",mem-quota-query = 1 << 30, and close copr-cache
User-veriables: tidb_hashagg_final_concurrency = 1; tidb_hashagg_partial_concurrency = 1;

```
Old Implement
[tidb]> desc analyze select avg(L_EXTENDEDPRICE) from lineitem group by L_PARTKEY   ,L_SUPPKEY , L_QUANTITY;
+---------------------------+-------------+----------+-----------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------+------+
| id                        | estRows     | actRows  | task      | access object  | execution info                                                                                                                                                                                                                                                                             | operator info                                                                                                                                                                                   | memory   | disk |
+---------------------------+-------------+----------+-----------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------+------+
| HashAgg_9                 | 14397287.20 | 16776228 | root      |                | time:4m23.7s, loops:16385                                                                                                                                                                                                                                                                  | group by:test.lineitem.l_partkey, test.lineitem.l_quantity, test.lineitem.l_suppkey, funcs:avg(Column#19, Column#20)->Column#18                                                                 | 3.14 GB  | N/A  |
| └─TableReader_10          | 14397287.20 | 17963641 | root      |                | time:3m40.2s, loops:41, cop_task: {num: 40, max: 32.5s, min: 1.55s, avg: 19.7s, p95: 31.7s, max_proc_keys: 613883, p95_proc_keys: 462536, tot_proc: 12m46.5s, tot_wait: 9.29s, rpc_num: 40, rpc_time: 13m6.7s, copr_cache: disabled}                                                       | data:HashAgg_5                                                                                                                                                                                  | 320.5 MB | N/A  |
|   └─HashAgg_5             | 14397287.20 | 17963641 | cop[tikv] |                | tikv_task:{proc max:19.4s, min:873ms, p80:18.1s, p95:19.3s, iters:17591, tasks:40}, scan_detail: {total_process_keys: 17996609, total_keys: 17996649, rocksdb: {delete_skipped_count: 0, key_skipped_count: 17996609, block: {cache_hit_count: 57170, read_count: 0, read_byte: 0 Bytes}}} | group by:test.lineitem.l_partkey, test.lineitem.l_quantity, test.lineitem.l_suppkey, funcs:count(test.lineitem.l_extendedprice)->Column#19, funcs:sum(test.lineitem.l_extendedprice)->Column#20 | N/A      | N/A  |
|     └─TableFullScan_8     | 17996609.00 | 17996609 | cop[tikv] | table:lineitem | tikv_task:{proc max:10.9s, min:511ms, p80:10.1s, p95:10.7s, iters:17591, tasks:40}                                                                                                                                                                                                         | keep order:false, stats:pseudo                                                                                                                                                                  | N/A      | N/A  |
+---------------------------+-------------+----------+-----------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------+------+
4 rows in set (4 min 24.836 sec)


New Implement:
[tidb]> desc analyze select avg(L_EXTENDEDPRICE) from lineitem group by L_PARTKEY   ,L_SUPPKEY , L_QUANTITY;
+---------------------------+-------------+----------+-----------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------+---------+
| id                        | estRows     | actRows  | task      | access object  | execution info                                                                                                                                                                                                                                                                             | operator info                                                                                                                                                                                   | memory   | disk    |
+---------------------------+-------------+----------+-----------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------+---------+
| HashAgg_9                 | 14397287.20 | 16776228 | root      |                | time:3m42.7s, loops:16385                                                                                                                                                                                                                                                                  | group by:test.lineitem.l_partkey, test.lineitem.l_quantity, test.lineitem.l_suppkey, funcs:avg(Column#19, Column#20)->Column#18                                                                 | 1.04 GB  | 2.80 GB |
| └─TableReader_10          | 14397287.20 | 17963641 | root      |                | time:58.2s, loops:41, cop_task: {num: 40, max: 35.4s, min: 1.64s, avg: 24s, p95: 33.5s, max_proc_keys: 613883, p95_proc_keys: 462536, tot_proc: 15m29.7s, tot_wait: 19.1s, rpc_num: 40, rpc_time: 16m0s, copr_cache: disabled}                                                             | data:HashAgg_5                                                                                                                                                                                  | 183.1 MB | N/A     |
|   └─HashAgg_5             | 14397287.20 | 17963641 | cop[tikv] |                | tikv_task:{proc max:19.5s, min:1.29s, p80:18.1s, p95:19.3s, iters:17591, tasks:40}, scan_detail: {total_process_keys: 17996609, total_keys: 17996649, rocksdb: {delete_skipped_count: 0, key_skipped_count: 17996609, block: {cache_hit_count: 57170, read_count: 0, read_byte: 0 Bytes}}} | group by:test.lineitem.l_partkey, test.lineitem.l_quantity, test.lineitem.l_suppkey, funcs:count(test.lineitem.l_extendedprice)->Column#19, funcs:sum(test.lineitem.l_extendedprice)->Column#20 | N/A      | N/A     |
|     └─TableFullScan_8     | 17996609.00 | 17996609 | cop[tikv] | table:lineitem | tikv_task:{proc max:11s, min:731ms, p80:10.1s, p95:10.8s, iters:17591, tasks:40}                                                                                                                                                                                                           | keep order:false, stats:pseudo                                                                                                                                                                  | N/A      | N/A     |
+---------------------------+-------------+----------+-----------+----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------+---------+
4 rows in set (3 min 43.035 sec)
```